:recycle: Standardize use travel authorization action logs.
